talentflow/services/user-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
# Use psycopg2-binary instead of asyncpg (pure Python driver, no compilation)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
email-validator>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
# Use kafka-python instead of aiokafka (pure Python)
kafka-python>=2.0.2
prometheus-fastapi-instrumentator>=6.1.0
python-multipart>=0.0.9
httpx>=0.27.0
